Move preference to main screen
