+++ app-apple/Sources/AppLibrary/Views/Preferences/PreferencesView.swift
@@ -55,6 +55,9 @@ public struct PreferencesView: View {
                 eraseCloudKitSection
             }
             NavigationLink(advancedTitle, destination: advancedView)
+            if isEnabled(.zigRuntime) {
+                cryptoBackendSection
+            }
         }
         .themeForm()
     }
@@ -65,6 +68,12 @@ private extension PreferencesView {
         nil,
         .light,
         .dark
+    ]
+
+    static let cryptoBackends: [CryptoBackend] = [
+        .openssl,
+        .mbedtls,
+        .native
     ]
 
     var systemAppearanceSection: some View {
@@ -146,11 +155,23 @@ private extension PreferencesView {
         Strings.Global.Nouns.advanced
     }
 
+    func isEnabled(_ flag: ABI.ConfigFlag) -> Bool {
+        !userPreferences.experimental.ignoredConfigFlags.contains(flag) &&
+        (configObservable.isActive(flag) || userPreferences.experimental.enabledConfigFlags.contains(flag))
+    }
+
+    var cryptoBackendSection: some View {
+        Picker(Strings.Views.Preferences.cryptoBackend, selection: userPreferences.binding(\.cryptoBackend)) {
+            Text(Strings.Global.Nouns.default).tag(0)
+            ForEach(Self.cryptoBackends, id: \.rawValue) {
+                Text($0.localizedDescription)
+            }
+        }
+        .themeContainerEntry()
+    }
+
     func advancedView() -> some View {
-        PreferencesAdvancedView(
-            userPreferences: userPreferences,
-            experimental: userPreferences.binding(\.experimental)
-        )
+        PreferencesAdvancedView(experimental: userPreferences.binding(\.experimental))
         .navigationTitle(advancedTitle)
     }
 }
